code 10: Body Condition Index catalog documentation.
