+++ src/main/java/com/project/app_dog_back/application/mapper/IIccMapper.java
@@ -1,6 +1,6 @@
 package com.project.app_dog_back.application.mapper;
 
-import com.project.app_dog_back.application.dto.IccDto;
+import com.project.app_dog_back.application.dto.catalog.IccDto;
 import com.project.app_dog_back.domain.model.catalog.Icc;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
